+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## v0.2.1
 - Include HttpRouter params into env params
+- Rename config method to settings
+- Include config into handler
 
 ## v0.2.0